+++ CHANGELOG.md
@@ -6,11 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
-## 2025.09.18
+## 2025.11.07
 
 ### Added
+
+- Designed abstract class for Secondary MusicPlaylist component
+
+### Updated
+
+- Changed design to include secondary methods including hasNoSongs(), shufflePlaylist(), sortPlaylist() (By title and artist), firstSong(), toString(), and equals().
+
+## 2025.09.18
 
 - Designed a Music Playlist component
 - Designed a Latin Library component
 - Designed a Crochet Project Tracker component
-